// Jenta/types.ts
export interface ImageFile {
  base64: string;
  mimeType: string;
  name: string;
}
// ...
export interface AudioFile {
  base64: string;
  name: string;
}